+++ src/main/java/dev/umc/healody/today/note/NoteService.java
@@ -4,8 +4,8 @@ import dev.umc.healody.today.note.dto.*;
 import dev.umc.healody.today.note.type.Hospital;
 import dev.umc.healody.today.note.type.Medicine;
 import dev.umc.healody.today.note.type.Symptom;
-import dev.umc.healody.user.User;
-import dev.umc.healody.user.UserRepository;
+import dev.umc.healody.usertest.User;
+import dev.umc.healody.usertest.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
