+++ src/main/java/dev/umc/healody/today/note/type/Hospital.java
@@ -1,7 +1,7 @@
 package dev.umc.healody.today.note.type;
 
 import dev.umc.healody.today.note.Note;
-import dev.umc.healody.user.User;
+import dev.umc.healody.usertest.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
