+++ src/main/java/dev/umc/healody/today/todo/Todo.java
@@ -1,6 +1,6 @@
 package dev.umc.healody.today.todo;
 
-import dev.umc.healody.user.User;
+import dev.umc.healody.usertest.User;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
