+++ src/main/java/dev/umc/healody/usertest/User.java
@@ -1,4 +1,4 @@
-package dev.umc.healody.user;
+package dev.umc.healody.usertest;
 
 import jakarta.persistence.*;
 import lombok.Getter;
